+++ output.txt
@@ -1,1 +1,2 @@
+inside minishell echo
 hello world
