+++ src/builtins/builtin_echo.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   builtin_echo.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: axlee <axlee@student.42.fr>                +#+  +:+       +#+        */
+/*   By: gyong-si <gyong-si@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/12 09:53:07 by axlee             #+#    #+#             */
-/*   Updated: 2024/05/31 14:18:25 by axlee            ###   ########.fr       */
+/*   Updated: 2024/06/02 12:57:26 by gyong-si         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,6 +14,7 @@
 
 void minishell_echo(t_shell *minishell)
 {
+	printf("inside minishell echo\n");
     int newline = 1;  // Default is to print newline at the end
 
     if (minishell->cmd_list == NULL)
